+++ Quartz.IDE/Windows/MainWindow.xaml.cs
@@ -1,0 +1,44 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Reactive.Disposables;
+using System.Text;
+using System.Threading.Tasks;
+using System.Windows;
+using System.Windows.Controls;
+using System.Windows.Data;
+using System.Windows.Documents;
+using System.Windows.Input;
+using System.Windows.Media;
+using System.Windows.Media.Imaging;
+using System.Windows.Navigation;
+using System.Windows.Shapes;
+using Librarium.WPF.Windows;
+using Quartz.IDE.ViewModels;
+using ReactiveUI;
+
+namespace Quartz.IDE.Windows
+{
+    /// <summary>
+    /// The main window for the program.
+    /// </summary>
+    public partial class MainWindow : BorderlessReactiveWindow<MainWindowViewModel>
+    {
+        public MainWindow()
+        {
+            this.InitializeComponent();
+            this.Closed += this.MainWindowClosed;
+            this.ViewModel = new MainWindowViewModel();
+
+            this.WhenActivated(dispose =>
+            {
+                this.OneWayBind(this.ViewModel,
+                    vm => vm.Title,
+                    view => view.Title)
+                    .DisposeWith(dispose);
+            });
+        }
+
+        private void MainWindowClosed(object sender, EventArgs args) => Application.Current.Shutdown();
+    }
+}